+++ CHANGELOG.markdown
@@ -7,6 +7,7 @@
 - fix error with tenant login when the tenant name was saved with capital letters in it
 - cleaned up license section of the Settings > About screen
 - load special error page instead of showing white-screen when running on PHP4
+- fix Settings > Twilio Account view for Connect tenants since the Token is not required
 
 
 ## OpenVBX 1.2.4